third_party/WebKit/LayoutTests/html/text_level_semantics/a-tooltip.html

<!DOCTYPE html>
<html>
<body onload="test()">
<script src="../../resources/js-test.js"></script>
<a id="target" href="#" title="This should appear as a tooltip">Hello.</a>
<script>
window.jsTestIsAsync = true;

function test() {
    if (window.eventSender) {
        eventSender.dragMode = false;
        eventSender.mouseMoveTo(target.offsetLeft, target.offsetTop + target.offsetHeight / 2);
        shouldBe("testRunner.tooltipText", "'This should appear as a tooltip'");

        target.style.display = "none";
        finishJSTest();
    }
}
</script>
</body>
</html>
