third_party/blink/web_tests/html/text_level_semantics/a-tooltip-expected.txt

PASS testRunner.tooltipText is 'This should appear as a tooltip'
PASS successfullyParsed is true

TEST COMPLETE
